#92 refactor: 순환참조 해결 위해 ChatNoticeService 생성
