+++ src/main/java/com/example/sharemind/chatMessage/application/ChatMessageService.java
@@ -1,7 +1,5 @@
 package com.example.sharemind.chatMessage.application;
 
-import com.example.sharemind.chat.domain.Chat;
-import com.example.sharemind.chatMessage.content.ChatMessageStatus;
 import com.example.sharemind.chatMessage.dto.request.ChatMessageCreateRequest;
 import com.example.sharemind.chatMessage.dto.response.ChatMessageGetResponse;
 
@@ -15,8 +13,4 @@ public interface ChatMessageService {
                            Boolean isCustomer);
 
     List<ChatMessageGetResponse> getChatMessage(Long chatId, Long messageId, Long customerId, Boolean isCustomer);
-
-    void createChatNoticeMessage(ChatMessageStatus chatMessageStatus, Chat chat);
-
-    void updateSendRequestMessageIsActivatedFalse(Chat chat);
 }
