+++ src/main/java/com/example/sharemind/chatMessage/application/ChatMessageServiceImpl.java
@@ -2,8 +2,6 @@ package com.example.sharemind.chatMessage.application;
 
 import com.example.sharemind.chat.application.ChatService;
 import com.example.sharemind.chat.domain.Chat;
-import com.example.sharemind.chat.exception.ChatErrorCode;
-import com.example.sharemind.chat.exception.ChatException;
 import com.example.sharemind.chatMessage.content.ChatMessageStatus;
 import com.example.sharemind.chatMessage.domain.ChatMessage;
 import com.example.sharemind.chatMessage.dto.request.ChatMessageCreateRequest;
@@ -11,8 +9,6 @@ import com.example.sharemind.chatMessage.dto.response.ChatMessageGetResponse;
 import com.example.sharemind.chatMessage.dto.response.ChatMessageWebSocketResponse;
 import com.example.sharemind.chatMessage.repository.ChatMessageRepository;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 import lombok.RequiredArgsConstructor;
@@ -66,28 +62,6 @@ public class ChatMessageServiceImpl implements ChatMessageService {
         Chat chat = chatService.getChatByChatId(chatId);
         chatMessageRepository.save(
                 chatMessageCreateRequest.toEntity(chat, isCustomer, ChatMessageStatus.MESSAGE));
-    }
-
-    @Transactional
-    @Override
-    public void createChatNoticeMessage(ChatMessageStatus chatMessageStatus, Chat chat) {
-        if (chatMessageStatus == ChatMessageStatus.SEND_REQUEST) {
-            ChatMessage chatMessage = new ChatMessage(chat, false, chat.getConsult().getCustomer().getNickname() + "님, 지금 바로 상담을 시작할까요?", chatMessageStatus);
-            chatMessageRepository.save(chatMessage);
-        } else if (chatMessageStatus == ChatMessageStatus.START) {
-            updateSendRequestMessageIsActivatedFalse(chat);
-            ChatMessage chatMessage = new ChatMessage(chat, false, "상담이 시작되었어요. \n" + LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy년 MM월 dd일 HH시 mm분 a")), chatMessageStatus);
-            chatMessageRepository.save(chatMessage);
-        }
-    }
-
-    @Transactional
-    @Override
-    public void updateSendRequestMessageIsActivatedFalse(Chat chat) {
-        ChatMessage chatMessage = chatMessageRepository.findByMessageStatusAndIsActivatedTrue(ChatMessageStatus.SEND_REQUEST);
-        if (chatMessage == null)
-            throw new ChatException(ChatErrorCode.CHAT_STATUS_NOT_FOUND, "채팅 요청 존재하지 않음");
-        chatMessage.updateIsActivatedFalse();
     }
 
     @Override
